refactor: resolve circular dependencies
